requirements.txt
flask
flask-cors
stripe
sqlalchemy
psycopg2-binary
python-dotenv
bcrypt
gunicorn
